# backend/src/main/java/com/library/lbms/controller/AuthController.java
package com.library.lbms.controller;

import com.library.lbms.dto.request.CreateUserRequest;
import com.library.lbms.dto.request.LoginRequest;
import com.library.lbms.dto.response.AuthResponse;
import com.library.lbms.dto.response.UserResponse;
import com.library.lbms.service.UserService;
import com.library.lbms.security.JwtService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.security.authentication.AuthenticationManager;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/v1/auth") 
@RequiredArgsConstructor
public class AuthController {

    private final UserService userService;
    private final AuthenticationManager authenticationManager;
    private final UserDetailsService userDetailsService;
    private final JwtService jwtService;

    @PostMapping("/register")
    public ResponseEntity<UserResponse> register(@RequestBody CreateUserRequest request) {
        return ResponseEntity.status(201).body(userService.createUser(request));
    }

    @PostMapping("/login")
public ResponseEntity<AuthResponse> login(@RequestBody LoginRequest request) {

    authenticationManager.authenticate(
            new UsernamePasswordAuthenticationToken(request.getEmail(), request.getPassword())
    );

    UserDetails userDetails = userDetailsService.loadUserByUsername(request.getEmail());
    String jwtToken = jwtService.generateToken(userDetails);
    userService.updateLastLogin(request.getEmail());

    return ResponseEntity.ok(AuthResponse.builder()
            .token(jwtToken)
            .type("Bearer")
            .build());
}
}